Refactor(map): Move `block.ts` `chunk.ts` and `dimension.ts` to `map/` folder
